Fix warning due to boost 1.66

+++ Release/include/pplx/threadpool.h
@@ -67,9 +67,10 @@ public:
     boost::asio::io_service& service() { return m_service; }
 
 protected:
-    threadpool(size_t num_threads) : m_service(num_threads) {}
+    threadpool(size_t num_threads) : m_service(static_cast<int>(num_threads)) {}
 
     boost::asio::io_service m_service;
 };
 
 }
+
